feat: add settings for sync
